+++ CHANGELOG.md
@@ -19,6 +19,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `binary_fields` property to `ResourceModel` to allow for automatic conversion of binary fields to UUIDs
 
+## [1.5.0] - 2026.03.05
+
+### Added
+
+- Added support for nested filters up to `max_related_depth`
+
+## [1.4.0] - 2026.02.24
+
+### Added
+
+- Added support for recursive joins in `ResourceModel`
+
 ## [1.3.1] - 2026.01.08
 
 ### Fixed